refactor: move credentials related files to 'src/auth/credentials'
