+++ src/auth/auth.ts
@@ -31,7 +31,7 @@ import { getResourceFromTreeNode } from '../shared/treeview/utils'
 import { Instance } from '../shared/utilities/typeConstructors'
 import { TreeNode } from '../shared/treeview/resourceTreeDataProvider'
 import { createInputBox } from '../shared/ui/inputPrompter'
-import { CredentialsSettings } from './credentialsUtilities'
+import { CredentialsSettings } from './credentials/credentialsUtilities'
 import { telemetry } from '../shared/telemetry/telemetry'
 import { createCommonButtons, createExitButton, createHelpButton, createRefreshButton } from '../shared/ui/buttons'
 import { getIdeProperties, isCloud9 } from '../shared/extensionUtilities'
@@ -44,10 +44,10 @@ import { createRegionPrompter } from '../shared/ui/common/region'
 import { SsoCredentialsProvider } from './providers/ssoCredentialsProvider'
 import { AsyncCollection, toCollection } from '../shared/utilities/asyncCollection'
 import { join, toStream } from '../shared/utilities/collectionUtils'
-import { getConfigFilename } from './sharedCredentialsFile'
-import { saveProfileToCredentials } from './sharedCredentials'
-import { SectionName, SharedCredentialsKeys, StaticProfile, StaticProfileKeyErrorMessage } from './types'
-import { throwOnInvalidCredentials } from './sharedCredentialsValidation'
+import { getConfigFilename } from './credentials/sharedCredentialsFile'
+import { saveProfileToCredentials } from './credentials/sharedCredentials'
+import { SectionName, SharedCredentialsKeys, StaticProfile, StaticProfileKeyErrorMessage } from './credentials/types'
+import { throwOnInvalidCredentials } from './credentials/sharedCredentialsValidation'
 import { TempCredentialProvider } from './providers/tempCredentialsProvider'
 
 export const ssoScope = 'sso:account:access'
