+++ src/auth/credentials/sharedCredentialsValidation.ts
@@ -8,9 +8,9 @@
 import * as nls from 'vscode-nls'
 const localize = nls.loadMessageBundle()
 import { CredentialsData, CredentialsKey, SectionName, SharedCredentialsKeys } from './types'
-import { ToolkitError } from '../shared/errors'
+import { ToolkitError } from '../../shared/errors'
 import { profileExists } from './sharedCredentials'
-import { getLogger } from '../shared/logger'
+import { getLogger } from '../../shared/logger'
 
 /** credentials keys and their associated error message, if they exists */
 type CredentialsErrors = CredentialsData
