+++ src/auth/wizards/createProfile.ts
@@ -13,9 +13,9 @@ import { createInputBox } from '../../shared/ui/inputPrompter'
 import { DefaultStsClient } from '../../shared/clients/stsClient'
 import { createCommonButtons } from '../../shared/ui/buttons'
 import { credentialHelpUrl } from '../../shared/constants'
-import { showLoginFailedMessage } from '../credentialsUtilities'
-import { ParsedIniData, Profile } from '../sharedCredentials'
-import { SharedCredentialsKeys } from '../types'
+import { showLoginFailedMessage } from '../credentials/credentialsUtilities'
+import { ParsedIniData, Profile } from '../credentials/sharedCredentials'
+import { SharedCredentialsKeys } from '../credentials/types'
 
 function createProfileNamePrompter(profiles: ParsedIniData) {
     return createInputBox({
