+++ src/test/credentials/sharedCredentials.test.ts
@@ -8,7 +8,7 @@ import * as path from 'path'
 import * as fs from 'fs-extra'
 import { EnvironmentVariables } from '../../shared/environmentVariables'
 import { makeTemporaryToolkitFolder } from '../../shared/filesystemUtilities'
-import { getCredentialsFilename, getConfigFilename } from '../../auth/sharedCredentialsFile'
+import { getCredentialsFilename, getConfigFilename } from '../../auth/credentials/sharedCredentialsFile'
 
 describe('sharedCredentials', function () {
     let tempFolder: string
