+++ src/test/shared/credentials/userCredentialsUtils.test.ts
@@ -11,7 +11,11 @@ import * as sinon from 'sinon'
 /* eslint @typescript-eslint/naming-convention: 0 */
 
 import { Uri } from 'vscode'
-import { getSectionDataOrThrow, loadSharedConfigFiles, mergeAndValidateSections } from '../../../auth/sharedCredentials'
+import {
+    getSectionDataOrThrow,
+    loadSharedConfigFiles,
+    mergeAndValidateSections,
+} from '../../../auth/credentials/sharedCredentials'
 import { UserCredentialsUtils } from '../../../shared/credentials/userCredentialsUtils'
 import { EnvironmentVariables } from '../../../shared/environmentVariables'
 import { makeTemporaryToolkitFolder } from '../../../shared/filesystemUtilities'
